services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.122.0",
    "faststream>=0.6.5",
    "fastui>=0.9.0",
    "httpx>=0.28.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "pytest==8.2.0",
    "pytest-asyncio>=1.3.0",
    "python-jose",
    "python-multipart>=0.0.20",
    "redis>=5.0.0",
    "setuptools>=80.9.0",
    "shared",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[tool.aerich]
tortoise_orm = "src.core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.uv.sources]
shared = { workspace = true }
